requirements.txt
transformers
streamlit
torch
pillow
torchvision